Strip all the parsers for a consolidated parser.
