+++ src/flag/tests/mod.rs
@@ -46,57 +46,3 @@ fn should_generate_correct_help_message_based_off_passed_arguments() {
         )
     );
 }
-
-#[test]
-fn should_match_parse_flags_that_match_store_true_actions() {
-    let input = "--version";
-    let short_input = "-v";
-    let flag = Flag::new()
-        .name("version")
-        .short_code("v")
-        .help_string("print command version")
-        .action(Action::StoreTrue);
-
-    assert_eq!(
-        Ok(MatchStatus::Match((
-            &input[input.len()..],
-            ("version".to_string(), Value::Bool(true))
-        ))),
-        flag.clone().parse(input)
-    );
-
-    assert_eq!(
-        Ok(MatchStatus::Match((
-            &input[input.len()..],
-            ("v".to_string(), Value::Bool(true))
-        ))),
-        flag.clone().parse(short_input)
-    );
-}
-
-#[test]
-fn should_match_parse_flags_that_match_store_false_actions() {
-    let input = "--no-ask";
-    let short_input = "-n";
-    let flag = Flag::new()
-        .name("no-ask")
-        .short_code("n")
-        .help_string("don't prompt user for input")
-        .action(Action::StoreFalse);
-
-    assert_eq!(
-        Ok(MatchStatus::Match((
-            &input[input.len()..],
-            ("no-ask".to_string(), Value::Bool(false))
-        ))),
-        flag.clone().parse(input)
-    );
-
-    assert_eq!(
-        Ok(MatchStatus::Match((
-            &input[input.len()..],
-            ("n".to_string(), Value::Bool(false))
-        ))),
-        flag.clone().parse(short_input)
-    );
-}
