+++ src/lib.rs
@@ -1,7 +1,9 @@
+use crate::flag::FlagOrValue;
 use crate::parsers::match_string;
 use parcel::prelude::v1::*;
 use parcel::ParseResult;
 use std::default;
+use std::env::Args;
 use std::fmt;
 
 mod flag;
@@ -75,17 +77,5 @@ impl default::Default for App {
             description: String::new(),
             version: String::new(),
         }
-    }
-}
-
-impl<'a> Parser<'a, &'a [&'a str], String> for App {
-    fn parse(&self, input: &'a [&'a str]) -> ParseResult<'a, &'a [&'a str], String> {
-        let next = &input[1..];
-        match_string(format!("{}{}", "--", "test".to_string()))
-            .parse(&input[0])
-            .map(|ms| match ms {
-                parcel::MatchStatus::Match((_, v)) => parcel::MatchStatus::Match((next, v)),
-                parcel::MatchStatus::NoMatch(n) => parcel::MatchStatus::NoMatch(&input[0..]),
-            })
     }
 }
